requirements.txt
datasets==2.13.1
sentence-transformers==2.2.2
transformers==4.42.0
torch>=1.12.0
weaviate-client==3.26.7
pinecone-client==2.2.3
pandas==2.2.2
numpy==1.26.2
scikit-learn==1.4.0
scipy==1.11.2
tqdm==4.66.1
seaborn==0.13.2
matplotlib==3.8.0
python-dotenv==1.0.0
requests==2.32.3
psutil==5.9.5
pytest==7.4.3
